+++ src/ir_RC5_RC6.cpp
@@ -10,11 +10,11 @@
 // Returns -1 for error (measured time interval is not a multiple of t1).
 //
 #if (DECODE_RC5 || DECODE_RC6)
-int getRClevel(decode_results *results, unsigned int *offset, int *used, int t1) {
-    int width;
+int getRClevel(decode_results *results, unsigned int *offset, uint8_t *used, int t1) {
+    unsigned int width;
     int val;
     int correction;
-    int avail;
+    uint8_t avail;
 
     if (*offset >= results->rawlen) {
         return SPACE;  // After end of recorded buffer, assume SPACE.
@@ -153,9 +153,9 @@ void IRsend::sendRC5ext(uint8_t addr, uint8_t cmd, boolean toggle) {
 //+=============================================================================
 #if DECODE_RC5
 bool IRrecv::decodeRC5() {
-    int nbits;
-    long data = 0;
-    int used = 0;
+    unsigned int nbits;
+    unsigned long data = 0;
+    uint8_t used = 0;
     unsigned int offset = 1;  // Skip gap space
 
     if (results.rawlen < MIN_RC5_SAMPLES + 2) {
@@ -234,7 +234,7 @@ void IRsend::sendRC6(uint32_t data, uint8_t nbits) {
     uint32_t mask = 1UL << (nbits - 1);
     for (uint8_t i = 1; mask; i++, mask >>= 1) {
         // The fourth bit we send is a "double width trailer bit"
-        int t = (i == 4) ? (RC6_T1 * 2) : (RC6_T1);
+        unsigned int t = (i == 4) ? (RC6_T1 * 2) : (RC6_T1);
         if (data & mask) {
             mark(t);
             space(t);
@@ -263,7 +263,7 @@ void IRsend::sendRC6(uint64_t data, uint8_t nbits) {
     uint64_t mask = 1ULL << (nbits - 1);
     for (uint8_t i = 1; mask; i++, mask >>= 1) {
         // The fourth bit we send is a "double width trailer bit"
-        int t = (i == 4) ? (RC6_T1 * 2) : (RC6_T1);
+        unsigned int t = (i == 4) ? (RC6_T1 * 2) : (RC6_T1);
         if (data & mask) {
             mark(t);
             space(t);
@@ -280,9 +280,9 @@ void IRsend::sendRC6(uint64_t data, uint8_t nbits) {
 //+=============================================================================
 #if DECODE_RC6
 bool IRrecv::decodeRC6() {
-    int nbits;
+    unsigned int nbits;
     uint32_t data = 0;
-    int used = 0;
+    uint8_t used = 0;
     unsigned int offset = 1;  // Skip first space
 
     if (results.rawlen < MIN_RC6_SAMPLES) {
